pyproject.toml
[tool.poetry]
name = "tableauapi"
version = "0.1.0"
description = "Connecting to the Tableau API to get meta-data of artifacts and maybe even more"
authors = ["Gal Polak <gal_polak@yahoo.com>"]
readme = "README.md"
packages = [{include = "tableauapi"}]

[tool.poetry.dependencies]
python = "^3.11"
tableauserverclient = "^0.25.0"
boto3 = "^1.26.0"
click = "^8.0.0"
requests = "^2.28.0"
pydantic = "^2.0.0"
rich = "^13.0.0"
python-dotenv = "^1.0.0"
typer = "^0.9.0"
colorama = "^0.4.6"
tabulate = "^0.9.0"
openpyxl = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.0.0"
pre-commit = "^3.0.0"

[tool.poetry.scripts]
tableau-cli = "tableauapi.cli.main:main"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"